Switch to URL path based control for Back and Reset
